v1: Move legacy sources under v1/ directories
Place legacy headers under include/v1 and sources under src/v1. Colocate
EEPROMHandler, INA226, and PocketPDPinOut with their lib folders. Adjust
includes in main and dependent headers.
