+++ .github/pull_request_template.md
@@ -1,0 +1,25 @@
+## Summary
+<!-- One paragraph: what changed and why. Skip if commit message already says it. -->
+
+## Linked issues
+<!-- Fixes #123, Refs #456 — leave blank if none -->
+
+## Hardware tested
+- [ ] HW1_0
+- [ ] HW1_1
+- [ ] None (no hardware change)
+
+How tested:
+<!-- e.g. "flashed HW1_1, ran 5 charge cycles, INA226 readings within 2%" -->
+
+## Breaking change / migration
+<!-- Check any that apply, describe below. Leave unchecked if none. -->
+<!-- - [ ] User-visible behavior (UI, button mapping, menu) -->
+<!-- - [ ] Serial / USB-PD protocol -->
+<!-- - [ ] EEPROM layout or calibration constants -->
+<!-- - [ ] Public lib API (`lib/*` headers) -->
+
+Details:
+
+## Notes
+<!-- Follow-ups, known issues, anything reviewer should know -->
