+++ src/main.cpp
@@ -1,5 +1,5 @@
 #include <Arduino.h>
-#include <StateMachine.h>
+#include <v1/StateMachine.h>
 
 StateMachine statemachine;
 
@@ -11,4 +11,3 @@ void loop()
 {
     statemachine.update();
 }
-
